+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to the "vscode-css2stylus" extension will be documented in t
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [0.0.1] - 2020-03-12
+## [1.1.0] - 2020-03-12
+### Added
+- Command `Convert Stylus to CSS`
+- Keyboard shortcut ctrl+alt+shift+c
+### Changed
+- Command name from `css2stylus: format selection` to `Convert CSS to stylus`
+
+## [1.0.0] - 2020-03-12
 ### Added
 - Initial work